выносит в отдельные компоненты «Список отзывов» и «Отзыв»
